+++ slippi/Core/Notifications/notifications.c
@@ -1,7 +1,7 @@
 #ifndef SLIPPI_CORE_NOTIFICATION_C
 #define SLIPPI_CORE_NOTIFICATION_C
 #include "notifications.h"
-#include "../../Chat/chat_notifications.c"
+#include "Chat/chat_notifications.c"
 
 GOBJ* _notificationsGOBJ = NULL;
 int LastNotificationMessageID = -1; // Initial Notification Message ID to always add to the bottom of the list
